divide title component
