+++ src/components/Buttons.jsx
@@ -1,14 +1,20 @@
 import React from 'react';
 import Button from './Button';
 
+const NUMBERS = [1, 2, 3, 4, 5];
+
 function Buttons({ onClickNumber }) {
+  function renderNumbers(number) {
+    return (
+      <Button key={number} onClick={() => onClickNumber(number)}>
+        {number}
+      </Button>
+    );
+  }
+
   return (
     <p>
-      {[1, 2, 3, 4, 5].map((i) => (
-        <Button key={i} onClick={() => onClickNumber(i)}>
-          {i}
-        </Button>
-      ))}
+      {NUMBERS.map(renderNumbers)}
     </p>
   );
 }
