+++ src/components/Page.jsx
@@ -2,11 +2,14 @@ import React from 'react';
 
 import Counter from './Counter';
 import Buttons from './Buttons';
+import Title from './Title';
+
+const TITLE = 'Counter';
 
 function Page({ count, onClickCounter, onClickNumber }) {
   return (
     <div>
-      <p>Counter</p>
+      <Title title={TITLE} />
       <Counter
         count={count}
         onClick={onClickCounter}
