+++ src/components/Title.jsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Title({ title }) {
+  return <p>{title}</p>;
+}
+
+export default Title;
